+++ src/interfaces/errors/BindingResolutionErrorInterface.ts
@@ -1,0 +1,3 @@
+export interface BindingResolutionErrorInterface extends Error {
+
+}